hotel_booking_system/hotel_search_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
SQLAlchemy-Utils
python-dotenv
pyjwt
passlib[bcrypt]
psycopg2-binary
